main

+++ api/creators.ts
@@ -1,14 +1,14 @@
 import express from "express";
 import { conn } from "../dbconnect";
-import { PersonData } from "../model/persondata";
-import mysql from "mysql";
+
 
 export const router = express.Router();
 router.use(express.json());
 
-router.post("/insert/:movid", (req, res) => {
-    const persondata: PersonData = req.body;
+router.post("/insert/:movid/:perid", (req, res) => {
     const movieId = req.params.movid;
+    const personId = req.params.perid;
+
 
     // ตรวจสอบว่าหนังที่เลือกมามีอยู่ในฐานข้อมูลหรือไม่
     let checkMovieSql = "SELECT * FROM `movie` WHERE `mid` = ?";
@@ -22,39 +22,43 @@ router.post("/insert/:movid", (req, res) => {
             return res.status(404).json({ success: false, message: 'Movie not found' });
         }
 
-        // หากพบหนัง
-        const { Name, Gender, Birth, img } = persondata;
-
-        // เพิ่มข้อมูลบุคคลลงในตาราง person
-        const insertPerson = 'INSERT INTO `person`(`Name`, `Gender`, `Birth`, `img`) VALUES (?, ?, ?, ?)';
-        conn.query(insertPerson, [Name, Gender, Birth, img], (error, personResult) => {
-            if (error) {
-                return res.status(500).json({ error: 'Error inserting Person' });
+        let checkPersonSql = "SELECT * FROM `person` WHERE `pid` = ?";
+        conn.query(checkPersonSql, [personId], (err, result) => {
+            if (err) {
+                throw err;
             }
 
-            // หา ID ของบุคคลที่เพิ่มเข้าไป
-            const personId = personResult.insertId;
+            // ถ้าไม่พบคน
+            if (result.length === 0) {
+                return res.status(404).json({ success: false, message: 'Person not found' });
+            }
 
-            // เพิ่มการเชื่อมโยงระหว่างหนังและนักแสดงในตาราง stars
-            let insertStars = 'INSERT INTO `creators`(`mid_cre_fk`, `pid_cre_fk`) VALUES (?, ?)';
-            insertStars = mysql.format(insertStars, [movieId, personId]);
 
-            conn.query(insertStars, (err, result) => {
-                if (err) {
-                    throw err;
+            const starsdata = {
+                mid_cre_fk: movieId,
+                pid_cre_fk: personId
+            };
+
+            const insertStars = 'INSERT INTO `creators`(`mid_cre_fk`, `pid_cre_fk`) VALUES (?, ?)';
+            conn.query(
+                insertStars,
+                [starsdata.mid_cre_fk, starsdata.pid_cre_fk],
+                (error, result) => {
+                    if (error) {
+                        return res.status(500).json({ error: 'Error inserting Person' });
+                    }
+                    res.status(201).json({ success: true, message: 'Insert Star(Person) Successful' });
                 }
+            );
 
-                // ส่งข้อมูลกลับไปยังผู้ใช้งาน
-                res.status(201).json({ success: true, message: 'Insert Stars(Person) Successful' });
-            });
         });
     });
 });
 
 
-router.delete("/insert/:movid/:strid", (req, res) => {
+router.delete("/delete/:movid/:perid", (req, res) => {
     const movieId = req.params.movid;
-    const starsId = req.params.strid;
+    const personId = req.params.perid;
 
 
     // ตรวจสอบว่าหนังที่เลือกมามีอยู่ในฐานข้อมูลหรือไม่
@@ -69,14 +73,28 @@ router.delete("/insert/:movid/:strid", (req, res) => {
             return res.status(404).json({ success: false, message: 'Movie not found' });
         }
 
-        const deleteStars = 'DELETE FROM `creators` WHERE mid_cre_fk = ? AND pid_cre_fk =?';
 
-        conn.query(deleteStars, [movieId, starsId], (err, result) => {
+
+        let checkPersonSql = "SELECT * FROM `person` WHERE `pid` = ?";
+        conn.query(checkPersonSql, [personId], (err, result) => {
             if (err) {
                 throw err;
             }
 
-            res.status(200).json({ success: true, message: 'Delete Star(Person) Successful' });
+            // ถ้าไม่พบคน
+            if (result.length === 0) {
+                return res.status(404).json({ success: false, message: 'Person not found' });
+            }
+
+            const deleteStars = 'DELETE FROM `creators` WHERE mid_cre_fk = ? AND pid_cre_fk =?';
+            conn.query(deleteStars, [movieId, personId], (err, result) => {
+                if (err) {
+                    throw err;
+                }
+
+                res.status(200).json({ success: true, message: 'Delete Star(Person) Successful' });
+            });
+
         });
     });
 });
